data_collection/requirements.txt
requests == 2.27.1
newspaper3k == 0.2.8